+++ patches/src/main/kotlin/app/revanced/patches/reddit/customclients/sync/syncforreddit/fix/slink/FixSLinksPatch.kt
@@ -6,8 +6,6 @@ import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.util.smali.ExternalLabel
 import app.revanced.patches.reddit.customclients.RESOLVE_S_LINK_METHOD
 import app.revanced.patches.reddit.customclients.SET_ACCESS_TOKEN_METHOD
-import app.revanced.patches.reddit.customclients.boostforreddit.fix.slink.getOAuthAccessTokenFingerprint
-import app.revanced.patches.reddit.customclients.boostforreddit.fix.slink.handleNavigationFingerprint
 import app.revanced.patches.reddit.customclients.fixSLinksPatch
 import app.revanced.patches.reddit.customclients.sync.syncforreddit.extension.sharedExtensionPatch
 
@@ -26,7 +24,7 @@ val fixSLinksPatch = fixSLinksPatch(
     execute {
         // region Patch navigation handler.
 
-        handleNavigationFingerprint.method.apply {
+        linkHelperOpenLinkFingerprint.method.apply {
             val urlRegister = "p3"
             val tempRegister = "v2"
 
@@ -46,7 +44,7 @@ val fixSLinksPatch = fixSLinksPatch(
 
         // region Patch set access token.
 
-        getOAuthAccessTokenFingerprint.method.addInstruction(
+        setAuthorizationHeaderFingerprint.method.addInstruction(
             0,
             "invoke-static { p0 }, $EXTENSION_CLASS_DESCRIPTOR->$SET_ACCESS_TOKEN_METHOD",
         )